+++ tests/AD-roles.test.js
@@ -62,6 +62,18 @@ test('getADRoleCred: the result is an AD role credential information retrieved',
 	return expect(data).toBeDefined();
 });
 
+test('rotateADRoleCred: the result is an AD role credential rotated', async () => {
+  const data = await vault.rotateADRoleCred(RootToken, RoleName);
+    console.log(data);
+	return expect(data).toBeDefined();
+});
+
+test('getADRoleCred: the result is an AD role credential information retrieved', async () => {
+  const data = await vault.getADRoleCred(RootToken, RoleName);
+    console.log(data);
+	return expect(data).toBeDefined();
+});
+
 test('listADRoles: the result is the AD roles listed', async () => {
   const data = await vault.listADRoles(RootToken);
     console.log(data);